Satoshi
Satoshi teamed up with the State Of California. A review of all Cryptocurrency apps will be taking place in order to Gurantee a Honest Network. There has been those who have attempted to suppress and undermine the efforts of the legend and the White Papper behind Bitcoin. Justice will be served....